+++ JustSlidin/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="ync.ysc3232.week7">
+    package="com.swe.justslidin">
 
     <application
         android:allowBackup="true"
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Week7">
+        android:theme="@style/Theme.JustSlidin">
         <activity
             android:name=".MainActivity"
             android:exported="true">
