+++ JustSlidin/app/src/main/java/com/swe/justslidin/io/Action.java
@@ -1,0 +1,4 @@
+package com.swe.justslidin.io;
+
+public interface Action {
+}
