+++ JustSlidin/app/src/main/java/com/swe/justslidin/models/Character.java
@@ -1,22 +1,27 @@
-package ync.ysc3232.week7.models;
+package com.swe.justslidin.models;
 
-public class Ball {
+public class Character {
+
     private float rad;
     private Position pos;
 
-    public Ball(float x, float y) {
+    public Character(float x, float y) {
         this.pos = new Position(x,y);
         this.rad = 30.0f; // TODO: Move the constant somewhere else!!!
     }
+
     public void move(Motion m) {
         this.pos.add(m);
     }
+
     public Position getPosition() {
         return this.pos;
     }
+
     public float getRadius() {
         return this.rad;
     }
+
     @Override
     public String toString() {
         return "Ball{" +
@@ -24,4 +29,5 @@ public class Ball {
                 ", pos=" + pos +
                 '}';
     }
+
 }
