+++ JustSlidin/app/src/main/java/com/swe/justslidin/models/Motion.java
@@ -1,6 +1,7 @@
-package ync.ysc3232.week7.models;
+package com.swe.justslidin.models;
 
 public class Motion extends Position {
+
     public Motion(float x, float y) {
         super(x, y);
     }
