--- JustSlidin/app/src/main/java/ync/ysc3232/week7/io/Action.java
@@ -1,4 +1,0 @@
-package ync.ysc3232.week7.io;
-
-public interface Action {
-}
