+++ JustSlidin/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">week7</string>
+    <string name="app_name">Just Slidin</string>
 </resources>
