+++ JustSlidin/settings.gradle
@@ -6,5 +6,5 @@ dependencyResolutionManagement {
         jcenter() // Warning: this repository is going to shut down soon
     }
 }
-rootProject.name = "week7"
+rootProject.name = "Just Slidin"
 include ':app'
